requirements.txt
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
icalendar
discord_webhook
winshell
pywin32
requests
requests-oauthlib
beatifulsoup4
stringcase
python-dateutil
tzlocal
pytz